+++ cross/auto_parameters/clean_data/__init__.py
@@ -1,1 +1,2 @@
 from .missing_values_handler import MissingValuesParamCalculator
+from .missing_values_indicator import MissingValuesIndicatorParamCalculator