+++ cross/transformations/__init__.py
@@ -1,5 +1,5 @@
 from .categorical_features import CategoricalEncoding
-from .clean_data import MissingValuesHandler
+from .clean_data import MissingValuesHandler, MissingValuesIndicator
 from .datetime_features import DateTimeTransformer
 from .features_reduction import ColumnSelection, DimensionalityReduction
 from .numerical_features import (