+++ cross/transformations/clean_data/__init__.py
@@ -1,1 +1,2 @@
 from .missing_values_handler import MissingValuesHandler
+from .missing_values_indicator import MissingValuesIndicator